back-end/requirements.txt
# FastAPI & Uvicorn
fastapi
uvicorn
pydantic
pydantic-settings

# LangChain ecosystem
langchain
langchain-google-genai
google-generativeai
langchain-community
langsmith

# Vector Database
faiss-cpu

# OCR
easyocr

# Search & Web
tavily-python
requests

# Utils
python-dotenv
pillow
numpy
pandas

# Document Processing
PyPDF2
python-multipart

# Optional: For better embeddings
sentence-transformers
huggingface-hub
